lowers adventurer cap to 10 (#1078)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
WOE BE UPON ADVENTURER MAINS GO BACK TO THE WANDERER QUEUE
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

![image](https://github.com/user-attachments/assets/c2911b66-507d-483e-9dd1-4b63dae31ce0)
someone on discord asked for it
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->