Bladesinger Improvement (#1125)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
For a rare role, this role is pretty hot garbage. This aims to improve
it.
Increases their starting spellpoints by one, and buffs their starting
stats.
at present, they have a -2 to Endurance, which is /hilariously/ bad for
ANY role expected in melee. And the bladesinger is meant to be a melee
role. Hell, even mages don't have a -2 in END.
I've given them a total buff of 6 trait points from their 3. Currently,
pally's get 5, and warriors get 7.
Edit: per feedback increased sword skill to 3, so mages can't just hit a
dummy and be on the same level as the bladesinger, and removed axe skill
to compensate.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
In truth, the bladesinger role is pretty terrible. It's probably the
worst role for melee, due to it's endurance, and it's limited spell
casting means it can't even do that. I get wanting to have a 'spread
thin, so they're not great at either melee or magic', but the stat
penalty was a bit too much.
increasing them up to 6, with 2 INT, 2 END, 1 con, and 1 speed feels
like a good in between, while still keeping the rare role appropriately
effective.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->