cheap swamp weed zigs and expensive love potions to merchant stock (#1100)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

adds both swampweed zigs (8 mammon) as well as love potions (600, may be
to low) to the merchant luxury tab.
## Why It's Good For The Game

the merchant has every other drug from the purity vender outside of
swamp weed and this has always disappointed me. the lowest swamp weed
from purity can be is 12 ( compared to normal zigs lows of 5 and 3 from
the merchant) also allows for some merchant selling normal zigs as swamp
weed zigs

the love potion package gives a alternative to getting a item thats
other wise locked behind playing a mage (shameful) ~or a hour of
dungeoneering~ at a price that hopefully should only mostly be seen by
towners!.