Warrior buff (#1085)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Buffs the warrior, now truly making it someone trained in all weapons
and being more accurate to their description.

General changes

- Now has steelhearted trait, no more stress from watching someone's
head fly off for the 123rd time.
- Changed CON from 2 to 1 since now they're actually skilled.

Warrior buffs
- Raised all weapon skills to Journeyman
- Now has a chance to spawn with different weapons, such as: Mace,
flail, and hatchet, to show their diversity with weapons.

Monster hunter buffs
- Raised crossbows/bows to 50/50 Journeyman/Expert
- Raised sneaking to Journeyman
- Raised swimming to Apprentice
- Raised climbing to Journeyman
- Raised crafting to Novice
- Raised knives to Journeyman

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
The warrior class was always supposed to be a "master of all arts of
combat" type of class, which it failed to be since it only had
journeyman on swords and no other reason to pick it as a class other
than the semi-higher stats (which then you'd probably just pick paladin,
which if we're being honest, are just better warriors right now).

The class is **very rarely** if ever used since usually other classes
are just simply better, war clerics had different stats but miracles and
expert maces (i'm going to change that in another PR), barbarians are
just stronger in stats and have crit resistance, paladins are warriors
but cooler and with miracles (might nerf it, tell me what you think),
etc.

Hopefully this'll make the class an actual option and properly skilled,
and maybe we'll start to see some warriors and slightly more diversity
in adventurer classes for fighting.