Pick intent now deals pick damage (#1139)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

The dagger's pick intent, as it is on the name and the attack in itself,
is supposed to deal pick damage. Someone forgot to change the damage
type from stab to pick, this fixes it.
Ever wondered why you can't pick with a dagger but can with a stake? Or
why pick had the same chance of hitting a crit when it *should* have a
much higher chance? This is the reason.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Makes the pick intent properly work now! Higher chance of artery crits
like it should, can now be used to pick rocks like the stake, all the
things it was supposed to do from the start.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->