Sewing Rework Atomization - Part 2 - Seamster Shop Rework (#1154)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

In this part 2 of the Sewing rework I'm reworking the Tailor shop where
the seamster works!

### I had 3 goals when making this:

1 - Improve lighting as the place was extremely dark before which made
it very hard to work in and made it feel abandoned.

2 - Giving it the Smithy treatment (since as mentioned in Part 1 #1151
that is the reference for the rework), making them function similarly by
moving the counter outside, adding a Vomitorium, shutters, a Peddler and
making it so clients can see the seamster working from outside.

3 - Making it more Home-y by adding food, moving some things around and
adding some decorations such as the fireplace and replacing the windows
with openable ones (for privacy).

4 - Giving the Seamster just a little bit more starting resources by
adding two shirts which may be salvaged by the mechanic introduced in
Part 1 #1151

5 - Making the seamster and their shop more sought after and entice RP,
this was achieved by adding a tanning rack (which the seamster usually
already builds) to support light armour crafting and expand the items a
seamster may offer to their clients, and by lighting the outside of the
shop, to capture player attention.

6 - Giving the Seamsters more space with which to do whatever. By moving
the counter to the street the Seamsters now have more space to build
racks or whatever. Up to them!

Client POV:

![image](https://github.com/user-attachments/assets/a18b61b2-7af2-4629-bb7b-4a8038a0d1c9)

The inside:

![image](https://github.com/user-attachments/assets/6c663ca4-5516-438e-ba73-d602cc1049d7)

Seamster POV:

![image](https://github.com/user-attachments/assets/999fd333-0fdc-4287-9089-f78284a117f4)

Closed Shop, making it absolutely obvious when the Seamster is and isn't
around:

![image](https://github.com/user-attachments/assets/84a14b9d-3c4c-48ab-b17c-7332032f542d)

~~A would be Thief approaches an open window:~~
Changed these into reinforced since we can't have nice things since the
monkeys will break them 😡

![image](https://github.com/user-attachments/assets/adfc2427-17af-40c0-98e0-1ab9523665b7)

StrongDMM image:

![image](https://github.com/user-attachments/assets/400bbd76-b6ce-43a8-889a-f2d2a296739b)

## Why It's Good For The Game

More power to the Seamstress, yada yada!
Should improve roleplay!
Also is part of my Sewing Rework so it strengthens Part 1 and coming
parts!

## Changelog

Add: Reworked the Seamster Shop
Add: Added a Tanning Rack, a Vomitorium, and other goodies to the
seamster shop, including crackers and water.
Add: Added more lights to the Seamster Shop, including one on the
outside.
Add: 2 White plain shirts will now spawn in the Seamster shop to be
salvaged for cloth.