Removes +1 Reading skill for Elves and Half Elves from their DNA. (#1146)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/7dd5f91a-ff4d-4d18-94d2-d2675ac33cef)
Apparently elves and half elves could always read regardless their role
because the skill was inherited due to their race.
This PR removes this.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

![image](https://github.com/user-attachments/assets/c00c8576-2fe9-43d2-ab85-8fc4d8217bfd)


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->