Changes to Stamina and Fatigue Calculation (#1138)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Hopefully this closes #1063 

While investigating this issue I found out that Fatigue is the Green bar
and Stamina is the Blue bar.

Of course, congratulations to me for discovering the obvious, but its
normal to want to want to swap those names.
In most games what you spend to attack is stamina. The one that
regenerates on itself after some time is stamina.
It would make sense if fatigue was the blue one, which is the one you
regain when you go to sleep. So... tiredness!

However that is not the case here, and changing that would be an
absurdly complex task, which would probably bring more bugs than
solutions.

Looking at the issue, as we see here:


![image](https://github.com/user-attachments/assets/8e50f0d0-6d1c-4534-a762-f1e4e88db3d2)

The names are indeed swapped, no biggy, we swap them back.

But regarding this...


![image](https://github.com/user-attachments/assets/7a9d8fd6-931f-46ed-ad2f-0da5afb66ec0)

I agree! I think its odd there is a calculation for Stamina and not one
for Fatigue, which is the one you actually need to attack!

So I added that calculation to fatigue!
Which takes your endurance, adds your athletics level into a nice bundle
and multiplies it. That's your Maximum Fatigue now!


![image](https://github.com/user-attachments/assets/3b46a7c5-2236-4443-b00e-bfa15ec6790e)

However, since I don't play mage classes, I kept stamina as it is! Being
affected by Endurance!

## Problems and Tribulations

1 - There's a tiny itty bitty chance the hud does not display this
correctly.

However, while testing I didn't notice anything odd and looking at how
the hud is updated in code it seems to be ready to do the math by
percentage, but you never know!

2 - I went back to Endurance at the request of the comments!

_2 - I'm still on the fence about INTELIGENCE being the determining
factor in stamina

I was considering CONDITION actually! I will need a Maints opinion on
this!_

## Why It's Good For The Game

This will probably strengthen those with an able body and who put some
time into leveling their athletics.
And weaken those who aren't so lucky!
Diversity is the spice of life!

## Changelog

Fix: Swapped Stamina and Fatigue hud names back into place
Code: Removed commented out code in rogfatstam.dm
Add: Max Fatigue is now calculated with Endurance and Athletics in mind