Pan to Skillet - Smeltable edition (#1163)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

There was a request:

![image](https://github.com/user-attachments/assets/3ac20803-c7c0-44dc-99db-37e4faeed5db)

While investigating I discovered that not only was it not smeltable, it
was actually made out of IRON.

And that its item name was steel pan, in the smith menu it was just
frypan which is the American version of the world.

So, in short.
Renamed the pan to Iron Skillet and the recipe to Skillet for
consistency and historical accuracy.

![image](https://github.com/user-attachments/assets/8159c8df-d99b-4748-98fc-75c19543a178)

![image](https://github.com/user-attachments/assets/5ffdb0fb-09c5-4ef9-9297-d67319fb5544)

Made it smeltable into iron.

![image](https://github.com/user-attachments/assets/3d1dc57a-59ce-4d53-ba9a-a6d4597d58a5)

That's about it! Simple PR.

## Why It's Good For The Game

Since we try out best to be historically accurate with some things I
think this goes in line with that.
Also tying lose ends of things that you expect to be smeltable but
aren't and removing the discrepancy between item names/descriptions and
the items that are required to create them.

## Changelog

tweak: Renamed "steel pan" to "iron skillet" 
tweak: Renamed the pan recipe from "Frypan" to "Skillet"
add: Skillets can now be smelted into Iron