Adds new Seed & Feed special, added Seelie language to Polyglot special. (#1164)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds exclusive special for Soilsons, Towners and Pilgrims. Weight is
100, so the same as Outlaw and Excommunicated.

![image](https://github.com/user-attachments/assets/098fa6f8-a916-4ae5-8081-9dd688ac5840)
What it does, +1 Wrestling, Stash with bag of 10 seeds, and a sickle.

![image](https://github.com/user-attachments/assets/cb180a51-b995-498a-b9f7-ab0a059665aa)
What seeds inside the bag.

![image](https://github.com/user-attachments/assets/a36c59cb-37e6-4e92-a3f4-1a709326dc4b)

Also added, Seelie language to Polyglot special.

![image](https://github.com/user-attachments/assets/5783b22f-1fd4-4a44-b52c-054b5b1a09ab)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Hoping this special would add more people to play peasant roles, and
having specials for less played roles is good.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->