Lupian Merc Update (#1069)
## About The Pull Request

Open for feedback.

Adds a new mercenary to the game. The ~~Black Star~~ ~~Silver Dawn~~
Forlorn Hope mercenary. With Vakra swept up in civil wars and the might
of the Amarir fractured, soldiers are left with splintered leadership
and oft nobody to follow, but the banner of the Forlorn Hope welcomes
them.

They were once a holy order, dedicated to the Ten and sworn to Noc, now
they welcome the wounded soldiers and homeless peasants as mercenaries
of faith and fortune. With so many taking up the banner, and their
already strained infrastructure spread thin, one can find these
mercenaries wearing the tatters of civil wars, huddled in the woods in
camps, swearing their service under dubious claims of righteous fervor
and honor. (Some are original members of the Holy Order. Some are
peasants wanting for shelter and food for their dying children. Some are
veteran soldiers, or those left behind, and know nothing other than a
life of soldiering.)

Spritework is courtesy of Bimmer (brigandine and volf helm), and Cre
(warhammer, heater shield and heraldry)

They also have their own combat music, give it a listen.


https://github.com/user-attachments/assets/081f309a-2781-4111-b3a5-5e924d8832a3

## Why It's Good For The Game

In my opinion mercenaries being restricted based on race and lore is a
GOOD thing. The main goal of this PR is to further expand on the
existing lore and to offer another side to mercenaries. With this PR
your mercenary can be explicitly religious, they could be a zealot, or
they could be a peasant with nowhere to turn to, or they could be a
soldier far from home who doesn't care for their religion and just needs
a job to keep them from the disgrace of banditry.