I thought I had fixed that yesterday. (#1182)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

More bugfixes on Malum Spells

For some reason, one of the fixes I'm pretty sure I pushed yesterday
wasn't in.

Detailed explanation : buyprice is the variable that will be compared to
the item prices to check if they can be chosen as a boon for the
crafter's covenant spell.
buyprice was setup as tithe (the value of the items sacrificed
multiplied by "divine_tax" a constant actually setup to 2 and used to
make everything more expensive (well if I was good at math)

As you may have realised already, the good equation is buyprice =
tithe/divine_tax. dividing the result of the sacrifice by two and as
such making everything more expensive than at the merchant.

## Why It's Good For The Game

Fixes good, yes yes.