Update stats.dm (#1196)
Buffs the stats of old people.

## About The Pull Request
This simple PR just gives old people +1 Fortune and changes intelligence
from +2 to +3


## Why It's Good For The Game

Elder characters are not often played. There are a few reasons for this
but I believe that one of the main reasons is because old people get
large minuses to almost every stat. The goal of this PR is simply to
give people a reason to play old characters. Greater intelligence allows
for higher skills (which old characters should always get because of
their age) and higher fortune rolls. I believe that elders are simply
lucky to live as long as they did on such a grim planet and that they
should get higher fortune for that reason.

---------

Co-authored-by: Moribun <117505981+Moribun@users.noreply.github.com>