neu food port (bounty) (#1202)
ports the new food system im... pretty sure @GeneralPantsuIsBadAtCoding
made. im really not sure on that its floated thru so many codebases that
I try to give credit where I can. stonehenge...stonehedge something was
mentioned with it too. either way its a port heavily modified. I tossed
in a few differences to fit here better and renamed the handpies to
Dobk-E-Pies as a head nod to the all powerful donkpockets. sugarcanes a
no, so I did beets for sugar, mill them. and I added a soup for
them...the beets.

added a limited amount of and massively overpriced seed vendor to the
soilsons and barkeep, cleaned up tavern BS via the bounty payers
reccomendation. and I added a few more purity stations around the map.
because I want you all to smoke more its good for you. also because my
products are in thos machines so its a conflict of interest but hay.....
smoke.