Merchant infinit butter money fix (#1209)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Corrects merchant's goldface price for eggs and butter as he can just
buy it there for cheap and sell to stockpile for higher price

Also fixes stockpile not havening poultry position as it was overridden
by eggs
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->