Gatemaster appreciation PR (#1213)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Hey dear gatemaster. 

I've realized the pain that your job is of sitting on a single place and
pulling levers and winches all day, so i've called the local masons and
made some modifications to your place to make your job slightly more
bearable.

Changes:

- Made the room slightly bigger, now you can comfortably have a second
person around if you'd like or fit even more ale barrels than you did
before.

- New table placed facing the manor, so now every time a maid wants to
hand you a pie they won't have to bother you by standing in the middle
of the gate and slowing everything down.

- New HERMES added to your station, sending letters or gifts to the king
or recieving new orders from the marshal has never been easier, since
now you don't even have to leave your post. Handy! (We had to change
most of the numbers of the other HERMES to accomodate this new one, but
we hope that the people can get used to this small change, they rarely
use them anyways.)

- Placed a shylock on your station so you can finally have a comfortable
way of getting change for those tolls

- Placed a chest with the same lock as your door, so feel free to lock
it with any weapons or contraband you don't want those pesky men-at-arms
stealing after you're gone for 3 minutes.

- I've gone through the old armory and found a custom-made helmet from
one of the first gatemasters that was left there collecting dust. I sent
it to the smiths for some repairs and cleaning, and i now give it to
you. Enjoy. (Below is a drawing of how we expect it to look when you're
wearing it. It fits so good specifically on you!)
![Captura de pantalla 2024-10-22
123003](https://github.com/user-attachments/assets/9e65590d-dcd8-4e0e-8aa9-b09bd245d3ae)

I hope this helps make the job slightly less torturous, and that it at
least stops you from dissapearing or throwing yourself into the pit.
Please. I'm tired of having to get a new gatemaster every other week,
and there aren't enough masochists to take the job.
--The marshal

Before: 
![Captura de pantalla 2024-10-22
122640](https://github.com/user-attachments/assets/30ef7be6-a278-4483-9c4d-205fabe81590)

After:
![Captura de pantalla 2024-10-22
122558](https://github.com/user-attachments/assets/43941045-2477-4437-9750-de4fa4a71045)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Makes playing gatemasters less of a pain! Hopefully it'll make slightly
more people play it. What else is there to say?
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->