Templar Lovin Pt.1 (#1221)
## About The Pull Request
- - -
Player Facing:
- Ravox Paladins, Templars and Clerics may now exist. The god of Justice
may now have his followers properly represented. We're just lacking
unique sprites. Same for Malum.
- Adds an initial spell for Ravox followers. Burden. Intended to allow
them to more easily catch someone fleeing. It'll probably need to be
tweaked. This is a copied tweak of the haste spell, turning it into a
debuff granter instead.
- - -
Other:
- Malum is now accounted for when it comes to Templars and Paladins,
although it grants nothing unique beyond the cross.
- Reordering of Templar and Paladin faith checks to make them match the
define.
 - Empty monk spell file removed.
- - -
## Why It's Good For The Game
It made no sense for this to not exist, and I figured I'd toss a spell
in, too.
Intended to add two other incantations along with it, but decided
against it for now. The descriptions are still there, if anyone wishes
to do it later, or just doing something different altogether.