Double bed craft (#1191)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes possible to craft double beds, both their variants

Changes nicebed craft result
on screenshot below from variant on the left to the one on the right

![image](https://github.com/user-attachments/assets/ba1def7a-6e3b-4d17-81ac-7b1fd149bfc5)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->