Don't cry (during sex), pussy (#1243)
## About The Pull Request


![5128262e0b9b923ae9bffbab493df792](https://github.com/user-attachments/assets/3a071e20-02c7-40ee-92e3-f7c9d9afce57)

Replaces the *cry emote during brutal sex for *groan instead.

## Why It's Good For The Game

Because it's annoying during brutal sexo if someone wants to be brutally
fucked. Let the hornies emote their tears instead.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->