Guillotine revisit  (#1261)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Guillotine now dont gives watched dismembering debuff, instead, if
execution was succesfull, crowd gets mood buff, and if it failed-
debuff, with executor getting failed execution mood debuff too.
Maximum level of guillotine sharpness was reduced for this system to
work better (ie executor needs to be caring about blade sharpness and
sharp it with stone if needed)

Removed code that let to fix (sharpen) guillotine with plasteel as it is
not being used.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->