Drips out the dungeoneer (#1264)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Replaces the dungeoneers belt to be black, gives them a black satchel
and a black cloak
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I am sick of running around roundstart for this stuff, all black
dungeoneer because edgy. The only buffs is the storage slot from a cloak
as well as the five from the satchel. This gives the dungeoneer more
time to babysit the prisoners and drips them the fuck out.

I like to use the satchel to also carry kidnapping supplies
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->