[TM First] Fix for Shirts clipping over armor (#1280)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Layering problem!
This hopefully fixes cases where shirts are rendering above armor on the
characters side view.

Here we can see the tunic no longer clips over the gamberson:

![image](https://github.com/user-attachments/assets/cb9db97d-4dcf-4d7d-b122-74ed78ba7458)

I recommend this isn't merged and TM'd instead, as I don't know if this
could have any unintended effects beyond what I've tested.

I'm currently also looking for a way to fix character markings not
appearing on the side view, I also believe that is a layering problem.

## Why It's Good For The Game

Quite an annoying bug this is!