Timed Gates - - - Simple timed gates to prevent bandits from rushing the town before thirty minutes is up. They take place as a 'fog', for which I don't have sprites. Anyone can pass through, with the exception of bandits, and none can see through them.
This is a simple five minute hack job, with accompanying map edits to prevent digging around them on the upper levels. You can still mine around by the water below the gate and further down, as I was told to keep that.

I'll be clear on the fact that I have not tested this. In theory, it should work, but I 'unno. We'll see, if this is TM'd.
- - -